cast arg_str_space to size_t when passed to strncat.
git-svn-id: svn+ssh://svn.nta-monitor.com/trunk/opensource/ike-scan@2310 062a1500-4a13-0410-a63b-ee65f32af78f

+++ ike-scan.c
@@ -186,7 +186,7 @@ main(int argc, char *argv[]) {
    for (arg=0; arg<argc; arg++) {
       arg_str_space -= strlen(argv[arg]);
       if (arg_str_space > 0) {
-         strncat(arg_str, argv[arg], arg_str_space);
+         strncat(arg_str, argv[arg], (size_t) arg_str_space);
          if (arg < (argc-1)) {
             if (--arg_str_space > 0) {
                strcat(arg_str, " ");
